move stride_funcs.h to kernels dir
